+++ bgce-timetracker/Controllers/TimesheetController.cs
@@ -17,32 +17,53 @@ namespace bgce_timetracker.Controllers
         // GET: Timesheet
         public ActionResult Index()
         {
-            var tIME_SHEET = db.TIME_SHEET.Include(t => t.PAY_PERIOD1).Include(t => t.USER).Include(t => t.USER1);
-            return View(tIME_SHEET.ToList());
+            if (Request.IsAuthenticated)
+            {
+                var tIME_SHEET = db.TIME_SHEET.Include(t => t.PAY_PERIOD1).Include(t => t.USER).Include(t => t.USER1);
+                return View(tIME_SHEET.ToList());
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // GET: Timesheet/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Request.IsAuthenticated)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                TIME_SHEET tIME_SHEET = db.TIME_SHEET.Find(id);
+                if (tIME_SHEET == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(tIME_SHEET);
             }
-            TIME_SHEET tIME_SHEET = db.TIME_SHEET.Find(id);
-            if (tIME_SHEET == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "Home");
             }
-            return View(tIME_SHEET);
         }
 
         // GET: Timesheet/Create
         public ActionResult Create()
         {
-            ViewBag.pay_period = new SelectList(db.PAY_PERIOD, "ppID", "created_by");
-            ViewBag.approved_by = new SelectList(db.USERs, "userID", "fname");
-            ViewBag.employee = new SelectList(db.USERs, "userID", "fname");
-            return View();
+            if (Request.IsAuthenticated)
+            {
+                ViewBag.pay_period = new SelectList(db.PAY_PERIOD, "ppID", "created_by");
+                ViewBag.approved_by = new SelectList(db.USERs, "userID", "fname");
+                ViewBag.employee = new SelectList(db.USERs, "userID", "fname");
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // POST: Timesheet/Create
@@ -52,35 +73,49 @@ namespace bgce_timetracker.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "timesheetID,employee,submitted_on,approved_on,approved_by,submitted,approved,comments,active,is_missing_punches,total_entries,total_hours_worked,total_overtime_worked,pay_period,total_pto_used,total_unpaid_time,created_on,total_pto_earned")] TIME_SHEET tIME_SHEET)
         {
-            if (ModelState.IsValid)
+            if (Request.IsAuthenticated)
             {
-                db.TIME_SHEET.Add(tIME_SHEET);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                if (ModelState.IsValid)
+                {
+                    db.TIME_SHEET.Add(tIME_SHEET);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
 
-            ViewBag.pay_period = new SelectList(db.PAY_PERIOD, "ppID", "created_by", tIME_SHEET.pay_period);
-            ViewBag.approved_by = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.approved_by);
-            ViewBag.employee = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.employee);
-            return View(tIME_SHEET);
+                ViewBag.pay_period = new SelectList(db.PAY_PERIOD, "ppID", "created_by", tIME_SHEET.pay_period);
+                ViewBag.approved_by = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.approved_by);
+                ViewBag.employee = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.employee);
+                return View(tIME_SHEET);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // GET: Timesheet/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Request.IsAuthenticated)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                TIME_SHEET tIME_SHEET = db.TIME_SHEET.Find(id);
+                if (tIME_SHEET == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.pay_period = new SelectList(db.PAY_PERIOD, "ppID", "created_by", tIME_SHEET.pay_period);
+                ViewBag.approved_by = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.approved_by);
+                ViewBag.employee = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.employee);
+                return View(tIME_SHEET);
             }
-            TIME_SHEET tIME_SHEET = db.TIME_SHEET.Find(id);
-            if (tIME_SHEET == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "Home");
             }
-            ViewBag.pay_period = new SelectList(db.PAY_PERIOD, "ppID", "created_by", tIME_SHEET.pay_period);
-            ViewBag.approved_by = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.approved_by);
-            ViewBag.employee = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.employee);
-            return View(tIME_SHEET);
         }
 
         // POST: Timesheet/Edit/5
@@ -90,31 +125,45 @@ namespace bgce_timetracker.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "timesheetID,employee,submitted_on,approved_on,approved_by,submitted,approved,comments,active,is_missing_punches,total_entries,total_hours_worked,total_overtime_worked,pay_period,total_pto_used,total_unpaid_time,created_on,total_pto_earned")] TIME_SHEET tIME_SHEET)
         {
-            if (ModelState.IsValid)
+            if (Request.IsAuthenticated)
             {
-                db.Entry(tIME_SHEET).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(tIME_SHEET).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.pay_period = new SelectList(db.PAY_PERIOD, "ppID", "created_by", tIME_SHEET.pay_period);
+                ViewBag.approved_by = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.approved_by);
+                ViewBag.employee = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.employee);
+                return View(tIME_SHEET);
             }
-            ViewBag.pay_period = new SelectList(db.PAY_PERIOD, "ppID", "created_by", tIME_SHEET.pay_period);
-            ViewBag.approved_by = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.approved_by);
-            ViewBag.employee = new SelectList(db.USERs, "userID", "fname", tIME_SHEET.employee);
-            return View(tIME_SHEET);
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // GET: Timesheet/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Request.IsAuthenticated)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                TIME_SHEET tIME_SHEET = db.TIME_SHEET.Find(id);
+                if (tIME_SHEET == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(tIME_SHEET);
             }
-            TIME_SHEET tIME_SHEET = db.TIME_SHEET.Find(id);
-            if (tIME_SHEET == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "Home");
             }
-            return View(tIME_SHEET);
         }
 
         // POST: Timesheet/Delete/5
@@ -122,10 +171,17 @@ namespace bgce_timetracker.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            TIME_SHEET tIME_SHEET = db.TIME_SHEET.Find(id);
-            db.TIME_SHEET.Remove(tIME_SHEET);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Request.IsAuthenticated)
+            {
+                TIME_SHEET tIME_SHEET = db.TIME_SHEET.Find(id);
+                db.TIME_SHEET.Remove(tIME_SHEET);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         protected override void Dispose(bool disposing)
